Only download binaryen if it is missing

+++ scripts/build.sh
@@ -55,10 +55,13 @@ wasm_build() {
         pushd $WASM_CRATE
         npm install
         popd
+        BINARYEN_MD5="cfc00c1f1a0c05c4c40a8bb222439c11"
         BINARYEN_URL="https://github.com/WebAssembly/binaryen/releases/download/1.38.26/binaryen-1.38.26-x86-linux.tar.gz"
         BINARYEN_DIR="binaryen-1.38.26"
-        curl -sSL $BINARYEN_URL -o - | tar xzv
         WASM2JS=$BINARYEN_DIR/wasm2js
+        if [[ ! -e $WASM2JS || $(md5sum $WASM2JS | cut -d' ' -f1) != $BINARYEN_MD5 ]]; then
+            curl -sSL $BINARYEN_URL -o - | tar xzv
+        fi
         $PYTHON $BUILD_WASM --manifest-path=$WASM_CRATE/Cargo.toml --webapp-dir=$WASM_CRATE \
             --target-dir=$TARGET --output-dir=$WEB_UPLOADS/$APP_NAME --branch=$TRAVIS_BRANCH --release
         $PYTHON $BUILD_WASM --manifest-path=$WASM_CRATE/Cargo.toml --webapp-dir=$WASM_CRATE \
